Reworking leap year calculation
I reworked the leap year calculation to be in accordance to the way it's calculated in the Gregorian system, in order to account for the date drifting caused by the Julian method.

Although this was not defined in the decrees instituting the republican calendar, it seems to be the accepted method for the usage of the republican calendar after year XIX, to prevent date drifting. [Wikipedia article](https://fr.wikipedia.org/wiki/Calendrier_r%C3%A9publicain#Ce_qu'aurait_pu_%C3%AAtre_l'%C3%A9volution_du_calendrier), for reference.